Included UART functionality for BBB. However, I can't tell if it is working on not because I suspect that the FTDI breakout board I have to connect the BBB to FRDM does not work correctly. Second, function pointers were created for decoding messages that are received.

+++ Modules/timers.c
@@ -66,7 +66,7 @@ void SetupPWM( uint8_t tpm, uint8_t ch, uint16_t modulus, uint8_t prescaler, uin
 //                                                                        *
 // Return Value:  NONE                                                    *
 //*************************************************************************
-void ChangeLEDPW( uint16_t pulseWidth )
+void ChangeLEDPW( uint8_t pulseWidth )
 {
    uint16_t valueToTPM0CnV = (uint16_t) TPM0_MOD * ( pulseWidth / 100.0 );
    uint16_t valueToTPM2CnV = (uint16_t) TPM2_MOD * ( pulseWidth / 100.0 );
